Move timeControls constants file to gameSetup feature directory
